# Reto2/Solucion/packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/contacts": {
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "displayName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "company": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "notes": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "emails": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                      "type": "object",
                      "required": [
                        "raw"
                      ],
                      "properties": {
                        "raw": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 254
                        },
                        "isPrincipal": {
                          "type": "boolean"
                        }
                      }
                    }
                  },
                  "phones": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                      "type": "object",
                      "required": [
                        "raw"
                      ],
                      "properties": {
                        "raw": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 32
                        },
                        "region": {
                          "type": "string",
                          "minLength": 2,
                          "maxLength": 2,
                          "description": "ISO 3166-1 alpha-2"
                        },
                        "isPrincipal": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateContactResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-01: crear contacto local",
        "tags": [
          "Contacts"
        ]
      },
      "get": {
        "operationId": "ContactsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "searchTerm",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListContactsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-11: buscar, filtrar y listar contactos",
        "tags": [
          "Contacts"
        ]
      }
    },
    "/v1/contacts/{id}": {
      "get": {
        "operationId": "ContactsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "UC-12: consultar detalle de contacto",
        "tags": [
          "Contacts"
        ]
      }
    }
  },
  "info": {
    "title": "SSOT Contacts API",
    "description": "Plataforma de Centralización de Contactos — Reto 2",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateContactResponse": {
        "type": "object",
        "properties": {
          "contactId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "contactId",
          "version"
        ]
      },
      "ContactSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "company": {
            "type": "string",
            "nullable": true
          },
          "primaryEmail": {
            "type": "string",
            "nullable": true
          },
          "primaryPhone": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "displayName",
          "company",
          "primaryEmail",
          "primaryPhone"
        ]
      },
      "ListContactsResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactSummaryResponse"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      }
    }
  }
}